LayoutTests/http/tests/navigation/ping-attribute/area-cross-origin-from-https-expected.txt

CONSOLE MESSAGE: [blocked] The page at https://127.0.0.1:8443/navigation/ping-attribute/area-cross-origin-from-https.html was not allowed to display insecure content from http://localhost:8000/navigation/resources/save-ping.py?test=/navigation/ping-attribute/area-cross-origin-from-https.html.

Ping not sent
